[feature] ultra short distance <1mile distance travel

+++ server/filters/filterGoogle.js
@@ -7,7 +7,7 @@ var filter = function(requestBody){
 
   //parse distance into an int
   distance = distance.replace(/\,/g,"").split(" ");
-  distance = parseInt(distance[0]);
+  distance = parseFloat(distance[0]);
 
   //The distance between each yelp query in miles (i.e. Yelp will be queried every 10 miles along the route)
   // if the distance is less than 20 miles, then query every total distance /10 miles to filter the waypoint
